requirements.txt
dash.ly==0.10.9
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.10.0.post2
pandas==0.16.2
plotly==1.8.6
python-dateutil==2.4.2
pytz==2015.6
requests==2.8.0
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
